fix: Endpoint "/v2/weather/{city}" error in model
