+++ Makefile
@@ -43,6 +43,10 @@ build: ## Build a beta version
 install: ## Install to $GOPATH/src
 	go install ./...
 
+run-local: ## Build a beta version
+	go build -o weather-api
+	./weather-api
+
 # Absolutely awesome: http://marmelab.com/blog/2016/02/29/auto-documented-makefile.html
 help:
 	@grep -E '^[a-zA-Z_-]+:.*?## .*$$' $(MAKEFILE_LIST) | awk 'BEGIN {FS = ":.*?## "}; {printf "\033[36m%-30s\033[0m %s\n", $$1, $$2}'
