+++ cache/manager_test.go
@@ -1,0 +1,1 @@
+package cache_test
